requirements.txt
python-chess
tensorflow
numpy
pandas
requests
fastapi
uvicorn
pydantic
python-multipart
tensorflow-metal
